commitlkzx

+++ ProyectoPSP7/app/src/main/java/com/aprendiz/ragp/proyectopsp7/models/AdapterCDY.java
@@ -46,7 +46,7 @@ public class AdapterCDY extends RecyclerView.Adapter<AdapterCDY.Holder>{
         public void connectData(Results results){
             txtPhase.setText(results.getPhase());
             txtTime.setText(Integer.toString(results.getTime()));
-            txtP.setText(Integer.toString(results.getPercent())+"%");
+            txtP.setText(Float.toString(results.getPercent())+"%");
         }
     }
 }
